html/browsers/browsing-the-web/navigating-across-documents/016-onclick-return-true-then-navigation.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<title>Link with onclick navigates iframe and returns true, which allows the href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.querySelector('iframe').src='call-parent-postmessage.html';reports.push('onclick');return true;" href="href.html">Test</a>

<script>
var test = async_test();
var reports = [];

window.onmessage = test.step_func(function(e) {
  reports.push(e.data);

  // "href" is expected last, if this is not received last or
  // not received at all, the test must fail.
  if (e.data === "href") {
    assert_equals(reports[0], "click");
    assert_equals(reports[1], "onclick");
    assert_equals(reports[2], "href");
    assert_equals(reports.length, 3);
  } else {
    throw new Error('navigating to href must abort not-yet-mature navigation');
  }
  test.done();
});

test.step(function() {
  reports.push("click");
  document.querySelector("a").click();
});
</script>

<!--
1. This test creates an iframe and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to navigate the iframe.
3. Navigating the iframe to the anchor's `href` must abort the not-yet-mature
    navigation that was started by the `onclick` handler.
4. If the semantics are implemented correctly, reports will be
    ["click", "onclick", "href"]

-->
